+++ packages/network-tracker/bin/tracker.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 /* eslint-disable @typescript-eslint/no-require-imports */
 const program = require('commander')
-const { startTracker } = require('..')
+const { startTracker } = require('../src/startTracker')
 const { Logger } = require('@streamr/utils')
 const ethers = require('ethers')
 const { SlackBot } = require('@streamr/slackbot')
